Re #8787: Fix deleting of blocks
EditableBody's delete did not set _p_changed, thus the changes were never actually committed. This wasn't noticed so far, since when you get the same app server thread on your next request, it still has those changes in memory.

add had the same issue, but it sets item.__name__, which "accidentally" sets _p_changed on the article, so it didn't even go wrong there, but... ugh.

I don't know whether it's worth writing a specific tests for this (that would have to operate with the working copy instead of maniuplating a "free-floating" article), but on the other hand, it went unnoticed for quite a while, so other tests didn't actually catch this. Humm.