Fixes #9922: Omit XML (and name) from the form for citations so they are saved properly
There even is a test for the form, but it didn't catch this bug:
Since the test sets up the XML for the block by assigning an empty string to an objectify tree,
this is serialized and deserialized by the formlib to yield a StringifiedElement.
As long as a StringifiedElement does not contain text children, its representation is u'',
so both the existing value (part of the XML tree of the article) and the form value
are deemed to be equal, so formlib does *not* overwrite the XML of the block.

In the real world, however, the citation block begins its life as an
empty element, is serialized/deserialized without any type hints, and
thus yields an ObjectifiedElement. So both values are *not* deemed
equal, and the block's XML is overwritten by the form value -- which
no longer is connnected to the Article's XML tree; all changes made to
it are not persisted.

To fix the test and make it meaningful again, we now create the block
element directly instead of assigning an empty string.