Re #11290: Only run JS if we're on an article editor page
We need to (once again) be careful with the event timing:
Droppers are set up on zeit.edit.editor:after-reload,
so registerHandler must happen before that. But to determine
on what kind of page we are, we need the editor-contents to already
have been loaded. Thus, we use window:script-loading-finished,
which happens at /just/ the right time.