Fixes #11730: Remember whether an image was set manually or by syncing
This enables more sync possibilities, e.g. removing the teaser image
also removes it from the body -- but only if the body image hasn't been changed manually.

The implementation with a separate flag on the block is a kludge, the
proper solution needs a real Reference object that could know whether
it was set automatically or manually, see re #10686.