re #12050: prevent further blur events before saving
When users where clicking *really* fast it was possible that the editable got
the focus again *before* the save method locked the editor. When they *then*
clicked fast enough outside the editor, another blur event is issued. The
blur handler would try to save the text again then, resutling in the error.